refactored and added onClick for text phrases
